[chore] move CUSTOMIZED_KERNEL get_device_bw patch into dynamicemb_util
Class-level monkey-patch on HardwarePerfConfig.get_device_bw, applied
once at dynamicemb_util module load alongside the other dynamicemb
patches. Drops the per-planner _estimator/_config private-attribute
walk and the hasattr assert in plan_util.py. Also tightens the comments
added in this PR to one line each.
