+++ tzrec/utils/plan_util.py
@@ -542,10 +542,6 @@ class EmbeddingStorageEstimator(ShardEstimator):
             return
 
         for sharding_option in sharding_options:
-            # ``module_type_key`` is precomputed on ShardingOption in
-            # torchrec 1.6 (types.py:1175); preferred over the legacy
-            # ``sharder_name(type(sharding_option.module[1]))`` shape that
-            # post-rc1 PR #3917 removes.
             sharder_key = sharding_option.module_type_key
             sharder = sharder_map[sharder_key]
 
@@ -788,60 +784,8 @@ class EmbeddingEnumerator(_EmbeddingEnumerator):
         global_constraints: Optional[ParameterConstraints] = None,
     ) -> None:
         if not estimator:
-            perf_estimator = EmbeddingPerfEstimator(
-                topology=topology, constraints=constraints
-            )
-            if has_dynamicemb:
-                # torchrec 1.6 raises ValueError from
-                # HardwarePerfConfig.get_device_bw when it sees a compute
-                # kernel that isn't in its internal kernel_bw_lookup table.
-                # dynamicemb's CUSTOMIZED_KERNEL is unknown to that lookup,
-                # so wrap the inner config's get_device_bw to compute its
-                # bandwidth from the per-shard caching_ratio (matching
-                # dynamicemb's original 1.5-era formula) and fall through to
-                # upstream for every other kernel. Using the method-level
-                # override instead of the static kernel_device_bandwidths
-                # dict preserves caching_ratio variation across the 10
-                # cache_load_factor copies emitted in `enumerate` — without
-                # it the perf model can't break ties on cache hit rate.
-                assert hasattr(perf_estimator, "_estimator") and hasattr(
-                    perf_estimator._estimator, "_config"
-                ), (
-                    "torchrec EmbeddingPerfEstimator no longer exposes "
-                    "_estimator/_config; tzrec dynamicemb integration "
-                    "needs a new hook."
-                )
-                _inner_cfg = perf_estimator._estimator._config
-                _orig_get_device_bw = _inner_cfg.get_device_bw
-
-                def _customized_kernel_aware_get_device_bw(
-                    compute_device: str,
-                    compute_kernel: str,
-                    hbm_mem_bw: float,
-                    ddr_mem_bw: float,
-                    ssd_mem_bw: float,
-                    hbm_to_ddr_mem_bw: float,
-                    caching_ratio: Optional[float] = None,
-                    prefetch_pipeline: bool = False,
-                ) -> Optional[float]:
-                    if compute_kernel == EmbeddingComputeKernel.CUSTOMIZED_KERNEL.value:
-                        cr = caching_ratio if caching_ratio is not None else 0.0
-                        return (cr * hbm_mem_bw + (1 - cr) * hbm_to_ddr_mem_bw) / 10
-                    return _orig_get_device_bw(
-                        compute_device,
-                        compute_kernel,
-                        hbm_mem_bw,
-                        ddr_mem_bw,
-                        ssd_mem_bw,
-                        hbm_to_ddr_mem_bw,
-                        caching_ratio,
-                        prefetch_pipeline,
-                    )
-
-                # pyre-ignore [8]
-                _inner_cfg.get_device_bw = _customized_kernel_aware_get_device_bw
             estimator: List[ShardEstimator] = [
-                perf_estimator,
+                EmbeddingPerfEstimator(topology=topology, constraints=constraints),
                 EmbeddingStorageEstimator(topology=topology, constraints=constraints),
             ]
         super().__init__(
@@ -879,12 +823,6 @@ class EmbeddingEnumerator(_EmbeddingEnumerator):
         self._sharder_map = {
             sharder_name(sharder.module_type): sharder for sharder in sharders
         }
-        # Forward-compat for torchrec ≥1.6.0 stable: post-rc1 commits
-        # b0027133 / 25b9b5ff (first tagged in v2026.03.30.00) wire
-        # ``_sharder_data_map`` into the enumerator and migrate every
-        # estimator/stat through it. v1.6.0-rc1 ignores this attribute, so
-        # populating it now is a cheap no-op that prevents the next torchrec
-        # bump from re-breaking the override.
         self._sharder_data_map = build_sharder_data_map(self._sharder_map)
         sharding_options: List[ShardingOption] = []
 
@@ -1030,15 +968,7 @@ class EmbeddingEnumerator(_EmbeddingEnumerator):
         return sharding_options
 
     def _get_num_buckets(self, parameter: str, module: nn.Module) -> Optional[int]:
-        """Get the per-table bucket count for virtual-table sharding.
-
-        Mirrors torchrec 1.6's ``EmbeddingEnumerator._get_num_buckets``
-        (enumerators.py:293-318): returns ``total_num_buckets`` when the
-        backing embedding config opts into virtual-table sharding,
-        otherwise ``None`` so the upstream
-        ``calculate_shard_sizes_and_offsets`` falls back to the standard
-        row-count split.
-        """
+        """Get total_num_buckets for virtual-table sharding, or None."""
         if isinstance(module, EmbeddingBagCollection):
             embedding_configs = module.embedding_bag_configs()
         elif isinstance(module, EmbeddingCollection):
@@ -1059,9 +989,7 @@ class EmbeddingEnumerator(_EmbeddingEnumerator):
         sharder_key: str = "",
     ) -> List[str]:
         _constraints, key = self._get_constraints(child_path, name)
-        # GRID_SHARD is only supported if specified by user in parameter constraints
-        # ROW based shardings on FeatureProcessedEmbeddingBagCollection require
-        # explicit user opt-in (matches torchrec 1.6 enumerators.py:344-365).
+        # GRID_SHARD and row-wise on FP modules require explicit opt-in.
         is_fp_module = "FeatureProcessedEmbeddingBagCollection" in sharder_key
         if not _constraints or not _constraints.get(key):
             filtered = [
